Testing to see how to add content to this site, and how to use Markdown
Trying to figure out how to add content to this site.  Also wanted
to see how to use Markdown.  Looks fairly simple, mainly a slight
change to the format of the top-of-page metadata.

This is an article I started to write quite some time ago.  Yes
it only covers iOS now, but I've always planned (and still plan)
on extending it to cover Android and other platforms, I just
haven't had a chance to get to it until now :) )