Add more to heartbleed article
This is even more draft than draft.  It was typed while I was out of
town on an iPad in various non-optimal environments, and contains lots
of spelling errors, typos, and stuff.  Alot of the sentences aren't
in their final form (more like a "stream of consciousness dump" that
needs fleshing out and ediitng)  In other words, Caveat Reader.